feat: Add navbar, footer, and video to all pages.
Add the same navbar and footer from 'index.html' to 'work.html', 'company.html',
and 'contact.html'. Add a customized  video for each of these particular
pages to their respective headers. Ensure that on mobile a customized
image is displayed in the header for each of these pages instead of the video.